+++ Source/Data/ChangelogData.lua
@@ -7,6 +7,7 @@ ChangelogData = {
         version = "0.8.0",
         changes = {
             { category = "Added", text = "Added Prospecting Info." },
+            { category = "Added", text = "Option to show or hide Expansion Information." },
             { category = "Changed", text = "Moved Expansion dropdown to Main Window header. Affects every tab." },
             { category = "Fixed", text = "Fixed label size in Options menu." },
         },